+++ src/app/search/(search)/SearchRank.tsx
@@ -2,7 +2,7 @@
 import { getRealtimeRank } from '@/factory/RealtimeRank';
 import React, { Dispatch } from 'react';
 import { RealtimeRankType } from '@/types/homeComponentsType';
-import SearchRankSkeleton from '../skeleton/SearchRankSkeleton';
+import SearchRankSkeleton from '../../../components/skeleton/SearchRankSkeleton';
 
 interface SearchRankType {
   setKeyWord: Dispatch<React.SetStateAction<string>>;
@@ -30,7 +30,11 @@ const SearchRank = ({ setKeyWord }: SearchRankType) => {
                 style={{ marginLeft: item.rank === 1 ? '0.12rem' : '0' }}>
                 {item.rank}위
               </div>
-              <div className="flex-1 text-body2 text-black truncate cursor-pointer" onClick={() => handleClick(item.keyword)}>{item.keyword}</div>
+              <div
+                className="flex-1 text-body2 text-black truncate cursor-pointer"
+                onClick={() => handleClick(item.keyword)}>
+                {item.keyword}
+              </div>
               {/* todo : 돋보기 누르면 그대로 검색 쿼리에 keyword가 들어가게하기 */}
               <div className="w-6 h-6 rounded-full bg-gray100 flex items-center justify-center cursor-pointer">
                 <img
