+++ src/app/search/(search)/SearchedResultItem.tsx
