+++ src/app/search/page.tsx
@@ -1,4 +1,4 @@
-import SearchIndex from '@/components/search/SearchIndex';
+import SearchIndex from '@/app/search/(search)/SearchIndex';
 import React from 'react';
 
 const SearchPage = () => {
